requirements.txt
fastapi
uvicorn
python-multipart
requests
jinja2
aiofiles
pywebview
pyinstaller
